OldCanary/requirements.txt
amqp==2.2.2
beautifulsoup4==4.6.0
cached-property==1.3.1
celery==4.1.0
dependency-injector==4.36.2
Django==1.11.11
django-crispy-forms==1.7.0
django-filter==1.1.0
django-tables2==1.14.2
django-tastypie==0.13.3
enum34==1.1.6
functools32==3.2.3.post2
gunicorn==19.7.1
html5lib==0.999999999
ipaddress==1.0.17
kombu==4.1.0
lxml==4.2.1
mock==2.0.0
numpy==1.13.3
pandas==0.22.0
Pillow==5.1.0
psycopg2-binary==2.7.4
python-dateutil==2.6.1
python-magic==0.4.17
python-memcached==1.53
python-mimeparse==0.1.4
pytz==2018.3
PyYAML==3.12
scipy==0.19.1
SecretStorage==2.3.1
semver==2.0.1
simplejson==3.13.2
six==1.11.0
sqlparse==0.2.4
subprocess32==3.2.7
tables==3.4.2
